services/image_tagger/requirements.txt
torch
transformers
pillow
open-clip-torch
fastapi
uvicorn 
python-multipart